Card Entity Refactor
- Splitting Card Dependencies
- Updating Card Repository
Refactoring Repositories

- Fixes compilation errors.
Avoids Erroring in Test


Gradle Update

- Updates Gradle to latest version [5.1.11].

+++ app/src/main/java/com/familyapps/cashflow/model/account/AccountRepository.kt
@@ -15,7 +15,7 @@ interface AccountRepository {
     fun updateAccount(account : Account)
 
     @Query("DELETE FROM ACCOUNT WHERE ACCOUNT_NUMBER = :accountNumber")
-    fun deleteAccount(accountNumber : String) : Long
+    fun deleteAccount(accountNumber : String) : Int
 
     @Query("SELECT * FROM ACCOUNT WHERE ACCOUNT_NUMBER = :accountNumber")
     fun findAccountByAccountNumber(accountNumber : String) : Account
